Add files via upload
This is the project written in the Java language. It is designed to analyze the weather from the file by displaying the number of weather data records, average low temperature, average high temperature, low temperature, high temperature, and precipitate on a specific date inputted by the user.
 There are different imports at the beginning of the code which enables the code to run correctly. For example, the java.io.IOException enable checking if there is an exception to be considered in the data which might cause an error, java.util.scanner is used to create a scanner object which is an extremely versatile object used to input alphanumeric characters from several input sources and convert them to binary data which can be read by the computer
The program starts with the scanner which is a class used to get input from the user. The scanner is enabled to work by java.util package which is one of the imports at the beginning of the program. The program then sets the date format which in our case follows the pattern of month(M), then day(d), then year(yyyy). After that, there is an Array list created to store weather data. An array is a container object that holds a fixed number of values of a single type. After that, the file name is declared. The name of the file can be changed here to match with the file the user has. If the filename is valid, the system reads each line from the file.
The program then creates the weather data list to read each column in the file as a unique type of data. The data in the file are separated by a comma, so, the system counts the data after every comma as unique. In every line, the first data before the first comma is declared as Date, the second data after the comma is declared as low temperature, the third line is declared as high temperature, and the fourth line is declared as precipitation.
The system then creates objects of data where it groups each line as one object containing four data, local date, low temperature, high temperature, and precipitation.
After dividing the data and creating different objects of data, the system starts interacting with the user by prompting him to enter the date he wants to analyze the weather data in a specific format. The program then compares the data entered to the local date in the file, it then converts the month entered as integers into the corresponding name of the month, for example, 1 is converted to January, 2 to February, and so on. In checking the month, the program uses a switch statement, every time a month is entered, the program starts to check from 1 which is January, once it finds the corresponding month, it breaks there and does not check the rest of the months. For example, if the month entered is April, the program checks if the number entered corresponds to January, then goes to February, until April, then breaks there and does not check from May to December.
After getting the date from the user, the system filters the weather date list to get the marching with the specified date. If the date format is wrong, the program gives an error and does not execute. Once the specified date is found in the file, the program output all the computed data corresponding to that date provided. In the output, the program includes the month and day which was initially put by the user, this is done to avoid confusion especially if the user wants to see the data for more than one day. The program terminates after giving the output, so the user needs to rerun if he wants to determine the weather analysis for another date.